(PC-27816)[PRO] refactor: move App hooks in App folder
