+++ pro/src/app/App/hook/__specs__/usePageTitle.spec.tsx
@@ -1,11 +1,11 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import React from 'react'
 import { Route, Routes } from 'react-router'
 import { Link } from 'react-router-dom'
 
-import usePageTitle from 'hooks/usePageTitle'
 import { renderWithProviders } from 'utils/renderWithProviders'
+
+import usePageTitle from '../usePageTitle'
 
 const PageTitle = (): null => {
   usePageTitle()
